(refactor) move auth logic from App.jsx to auth.js
