+++ client/src/js/App.jsx
@@ -1,9 +1,8 @@
 // This is the App Component, it hosts App-wide resources.
 
-var React = require('react/addons');
+var React = require('react');
 var Landscape = require('./components/Landscape.jsx');
-var keys = require('../../../server/config/secureAuth.js');
-var $ = require('jquery');
+var auth = require('./stores/auth.js');
 
 var App = React.createClass({
 
@@ -64,300 +63,10 @@ var App = React.createClass({
     return startOfDay;
   },
 
-  // This property holds all Authentication logic, it holds app and setAJAXParams in closure scope.
-  auth: function(){
-
-    // We set 'app' to 'this' so that we can access this.state within different contexts.    
-    var app = this;
-
-    // Set AJAXParams inputs a provider and task and returns an object which our AJAX calls use to set their options.
-    var setAJAXParams = function(provider, usage, param) {
-      var callLoc = provider + '-' + usage;
-      var updateState = function(update) {
-        app.setState(React.addons.update(app.state, update));
-      };
-
-      // This switch statement sets all properties necessary to make an AJAX call.  This allows us to create one AJAX call, and make different calls depending on provider.
-      switch(callLoc) {
-        case 'github-login':
-          callParams = {
-            url: 'https://github.com/login/oauth/authorize?client_id=' + keys.github.clientID,
-            callback: function(code) {
-              console.log('in login');
-              updateState({
-                userInfo: {github: {
-                  code: {$set: code}
-                }}
-              });
-            }
-          };
-          break;
-        case 'github-getToken':
-          callParams = {
-            url: 'http://localhost:8000/api/auth/getToken/',
-            data: {
-              code: param,
-              provider: 'github'
-              // client_id : keys.github.clientID,
-              // client_secret : keys.github.clientSecret
-            },
-            // redirect_uri: 'https://eihfnhkmggidbojcjcgdjpjkhlbhealk.chromiumapp.org/githubToken',
-            callback: function(res){
-              var token = res.match(/(?:access_token=)[a-zA-Z0-9]+/g)[0].split('access_token=')[1];
-              updateState({ 
-                userInfo: {github: {token: {$set: token} } } 
-              });
-              console.log('User info saved after login: ', app.state.userInfo);
-
-                // We need to refactor this call to work with all APIs
-              app.auth.makeRequest(provider, 'user'); 
-            }
-          };
-          break;
-        case 'github-user':
-          callParams = {
-            url: 'https://api.github.com/user',
-            data: {access_token: app.state.userInfo.github.token},
-            callback: function(user) {
-              updateState({
-                userInfo: {github: {
-                  name: {$set: user.name},
-                  username: {$set: user.login},
-                  reposUrl: {$set: user.repos_url}
-                }}
-              });
-
-              app.auth.makeRequest(provider, 'repos');
-            }
-          };
-          break;
-        case 'github-repos': 
-          callParams = {
-            url: app.state.userInfo.github.reposUrl,
-            data: {access_token: app.state.userInfo.github.token},
-            callback: function(repos){
-              var reposList = [];
-              
-              repos.forEach(function(repo) {
-                reposList.push(repo.name);
-              });
-              updateState({
-                userInfo: {github: {
-                  repos: {$set: reposList}
-                }}
-              });
-              // console.log('Saved user repos: ', reposList);
-              // console.log('Confirm via log User');
-
-              app.state.userInfo.github.repos.forEach(function(repo) {
-                app.auth.makeRequest('github', 'commits', repo);
-              });
-            }
-          };
-          break;
-        case 'github-commits':
-          callParams = {
-            url: 'https://api.github.com/repos/' + app.state.userInfo.github.username + '/' + param + '/commits?author=' + app.state.userInfo.github.username + '&since=' + app.state.day,
-            data: {access_token: app.state.userInfo.github.token},
-            callback: function(commits) {
-              commits.forEach(function(commitInfo) {
-                updateState({
-                  userInfo: {github: {
-                    commitsByRepo: {$push: [{repo: param, stats: commitInfo}]},
-                    totalCommits: {$set: app.state.userInfo.github.totalCommits + 1}
-                  }}
-                });
-              });
-            }
-          };
-          break;
-
-        case 'fitbit-login':
-          callParams = {
-            url: 'https://api.fitbit.com/oauth/request_token?oauth_callback=https://eihfnhkmggidbojcjcgdjpjkhlbhealk.chromiumapp.org/fitbit&oauth_consumer_key=' + keys.fitbit.consumerKey
-          };
-          break;
-
-        case 'jawbone-login':
-          callParams = {
-            url: 'https://jawbone.com/auth/oauth2/auth?response_type=code&client_id=' + keys.jawbone.clientID + '&scope=move_read&redirect_uri=https://eihfnhkmggidbojcjcgdjpjkhlbhealk.chromiumapp.org/jawbone',
-            callback: function(code) {
-              console.log('in login');
-              updateState({
-                userInfo: {fitness: {
-                  provider: {$set: provider},
-                  code: {$set: code}
-                }}
-              });
-            }
-          };
-          break;
-        case 'jawbone-getToken':
-          callParams = {
-            url: 'https://jawbone.com/auth/oauth2/token',
-            data: {
-              client_id: keys.jawbone.clientID,
-              client_secret: keys.jawbone.clientSecret,
-              grant_type: 'authorization_code',
-              code: param,
-            },
-            callback: function(res) {
-              var token = res.access_token;
-              app.setState(React.addons.update(app.state, {
-                userInfo: {fitness: {token: {$set: token} } }
-              }));
-              console.log(res);
-              console.log('saved userInfo: ', app.state.userInfo.fitness);
-              app.auth.makeRequest(provider, 'user');
-            },
-          };
-          break;
-        case 'jawbone-user':
-          callParams = {
-            url: 'https://jawbone.com/nudge/api/v.1.1/users/@me',
-            header: {'Authorization': 'Bearer ' + app.state.userInfo.fitness.token},  
-            callback: function(res){
-              updateState({
-                userInfo: {fitness: {
-                  firstName: {$set: res.data.first},
-                  lastName: {$set: res.data.last},
-                  xid: {$set: res.data.xid}
-                }}
-              });
-              app.auth.makeRequest('jawbone', 'moves');
-            }
-          };
-          break;
-        case 'jawbone-moves':
-          callParams = {
-            url: 'https://jawbone.com/nudge/api/v.1.1/users/@me/moves',
-            header: {'Authorization': 'Bearer ' + app.state.userInfo.fitness.token},  
-            callback: function(res){
-              console.log(res);
-              updateState({
-                userInfo: {fitness: {
-                  moves: {$set: res.data},
-                }}
-              });
-            }
-          };
-          break;
-        case 'paired-createAccount':
-        callParams = {
-          url: 'http://localhost:8000/api/auth/createAccount/',
-          data: {
-            accountCodes: param,
-          },
-          callback: function(res) {
-            console.log(res);
-            localStorage.setItem('xid', res.xid);
-            localStorage.setItem('commits', res.commits);
-            localStorage.setItem('steps', res.steps);
-          }
-        };
-        break;
-        // This case passes the userID from localStorage, and returns the user's info if there is a user corresponding to that xid.
-        case 'server-loadAccount':
-        callParams = {
-          url: 'http://localhost:8000/api/user/load',
-          data: {
-            xid: localStorage.xid,
-          },
-          callback: function(res) {
-            console.log('this is the response from our server, should be our account: ', res);
-          }
-        };
-        break;
-      }
-      return callParams;
-    };
-
-    return {
-      // This function is modularized to handle all Login requests for all APIs
-      login: function(provider) {
-        var callParams = setAJAXParams(provider, 'login');
-        // console.log('Ajax call with params: ', callParams); 
-
-        chrome.identity.launchWebAuthFlow({
-          'url': callParams.url,
-          'interactive': true
-          },
-          function(redirectUrl) {
-            var user = app.state.userInfo;
-            var code = redirectUrl.split('?code=')[1];
-            callParams.callback(code);
-            if( user.fitness.code === null || user.github.code === null ) {
-              console.log('authenticate with both providers please');
-            }
-          }
-        );
-      },
-
-      pairAccounts: function(){
-        var user = app.state.userInfo;
-        if( user.fitness.code === null || user.github.code === null ) {
-          console.log('authenticate with both providers please');
-        } else {
-          var accounts = {
-            github: {
-              code: user.github.code
-            },
-            fitness: {
-              provider: user.fitness.provider,
-              code: user.fitness.code
-            }
-          }
-          app.auth.makeRequest('paired', 'createAccount', accounts);
-        }
-      },
-
-      // Make a call to server to re-load the data being stored in the database.
-      loadServerAccount: function(){
-        app.auth.makeRequest('server', 'loadAccount');
-      },
-      
-      // This function is modularized to make all GET requests for all APIs
-      makeRequest: function(provider, usage, param) {
-        var callParams = setAJAXParams(provider, usage, param);
-        $.ajax({
-          type: 'GET',
-          url: callParams.url,
-          headers: callParams.header,
-          data: callParams.data,
-          success: function(res) {
-            callParams.callback(res);
-          },
-          fail: function(err) {
-            console.error('GET request failed: ', err);
-          }
-        });
-      },
-
-      // This function is modularized to make all POST requests for all APIs
-      postRequest: function(provider, usage, param) {
-        var callParams = setAJAXParams(provider, usage, param);
-        $.ajax({
-          type: 'POST',
-          url: callParams.url,
-          data: callParams.data,
-          redirect_uri: callParams.redirect_uri,
-          success: function(res) {
-            console.log('POST response: ', res);
-            callParams.callback(res);
-          },
-          fail: function(err) {
-            console.error('POST request failed: ', err);
-          }
-        });
-
-      }
-    };
-  },
-
   render: function() {
     return (
       <div id="landscape-container">
-        <Landscape userInfo={this.state.userInfo} auth={this.auth} />
+        <Landscape userInfo={this.state.userInfo} auth={auth} />
       </div>
     );
   }
